requirements.txt
# 核心框架
streamlit==1.31.0

# AI模型支持（按需选择）
anthropic==0.18.1  # Claude支持
openai>=1.0.0  # OpenAI GPT-4o支持

# 文档处理
PyMuPDF==1.23.22  # PDF处理（支持表格提取）
python-docx==1.1.0  # Word读写（支持Markdown转Word）
openpyxl==3.1.2  # Excel处理
pandas==2.2.0  # 数据处理

# OCR识别（扫描版PDF支持）
rapidocr-onnxruntime==1.3.22  # 轻量级OCR引擎

# 数据库
sqlalchemy==2.0.27

# 工具库
python-dotenv==1.0.1  # 环境变量管理
